Initial version of VMSS module
I've added the terraform definitions to create an Azure VMSS running a Spacelift private worker pool. The way it works is pretty similar to https://github.com/spacelift-io/terraform-aws-spacelift-workerpool-on-ec2, but just using an Azure VMSS instead of an AWS ASG.

One change that I've made compared to the AWS version is that I'm storing the startup script in the `/var/lib/cloud/scripts/per-boot` rather than passing it as the user_data directly. This means that the worker starts again if the instance is rebooted.