Allow the process exit behavior to be specified
When the VM is shutdown in Azure, it just stays in a turned off state, and Azure doesn't automatically replace it. We could probably setup a health check that failed if the spacelift process wasn't running, but for now I figured it's easier to just reboot the VM rather than turning it off.

I've added three options:

- Reboot - reboots the VM, which will cause the worker to be updated.
- Shutdown - shuts down the VM.
- None - does nothing. This is useful if there's a configuration issue, and you want to be able to connect to investigate.

I've also fixed a mistake in the usage info in the README file.