Fix issues with Bastion example
The `random_pet` declared in main.tf was a duplicate that I added by accident when moving the random ID to the resource group name.

I've removed the dependency on the `vmss_contributor` role assignment - there isn't much point since the VMs don't need that assignment to start, and I deliberately left the assignment commented out just as an example since it's not required.