feat: download correct binary when in FedRAMP environment (#26)
I've updated the startup script to check whether it's connecting to the MQTT broker for our FedRAMP environment. If so it downloads the FedRAMP-specific version of the launcher.

As part of this I've had to update the test cases to update the Azure provider and set the `resource_provider_registrations = "none"` option. This isn't related to the change itself, it's just that the tests began to fail presumably because of a change in Azure.